Web: Files: refactoring: moved hocs to HOCs folder, Badges, Icons, EditingWrapperComponents, EmptyContainer to components folder
