+++ products/ASC.Files/Client/src/HOCs/withLoader.js
