+++ products/ASC.Files/Client/src/components/EmptyContainer/EmptyContainer.js
