+++ products/ASC.Files/Client/src/components/EmptyContainer/EmptyFolderContainer.js
