+++ products/ASC.Files/Client/src/components/Icons.js
@@ -2,9 +2,9 @@ import styled from "styled-components";
 
 import commonIconsStyles from "@appserver/components/utils/common-icons-style";
 
-import FavoriteIcon from "../../../../../../../public/images/favorite.react.svg";
-import FileActionsConvertEditDocIcon from "../../../../../../../public/images/file.actions.convert.edit.doc.react.svg";
-import FileActionsLockedIcon from "../../../../../../../public/images/file.actions.locked.react.svg";
+import FavoriteIcon from "../../public/images/favorite.react.svg";
+import FileActionsConvertEditDocIcon from "../../public/images/file.actions.convert.edit.doc.react.svg";
+import FileActionsLockedIcon from "../../public/images/file.actions.locked.react.svg";
 
 export const EncryptedFileIcon = styled.div`
   background: url("images/security.svg") no-repeat 0 0 / 16px 16px transparent;
