+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/FilesRow/SimpleFilesRow.js
@@ -7,8 +7,8 @@ import FilesRowContent from "./FilesRowContent";
 import { withRouter } from "react-router-dom";
 import { createSelectable } from "react-selectable-fast";
 
-import withFileActions from "../hoc/withFileActions";
-import withContextOptions from "../hoc/withContextOptions";
+import withFileActions from "../../../../../../HOCs/withFileActions";
+import withContextOptions from "../../../../../../HOCs/withContextOptions";
 
 const StyledSimpleFilesRow = styled(Row)`
   margin-top: -2px;
